fix: сделать поиск событий по статусу регистронезависимым refactor: убрать в серисах зависимости от других сервисов style: fix

+++ ewm-service/src/main/java/ru/practicum/explorewme/request/RequestRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 public interface RequestRepository extends JpaRepository<Request, Long> {
     Optional<Request> findByRequesterIdAndEventId(Long userId, Long eventId);
 
-    Long countByEventIdAndStatus(Long eventId, RequestStatus status);
+    Integer countByEventIdAndStatus(Long eventId, RequestStatus status);
 
     List<Request> findAllByRequesterId(Long userId);
